% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.R
\name{ols}
\alias{ols}
\title{Ordinary least squares with generalized inverse from the ddml package
See ?ddml::ols}
\usage{
ols(y, X, const = FALSE, w = NULL)
}
\description{
Ordinary least squares with generalized inverse from the ddml package
See ?ddml::ols
}
