% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcmeans.R
\name{predict.kcmeans}
\alias{predict.kcmeans}
\title{Prediction Method for the K-Conditional-Means Estimator.}
\usage{
\method{predict}{kcmeans}(object, newdata, clusters = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{kcmeans}.}

\item{newdata}{A (sparse) feature matrix where the first column corresponds
to the categorical predictor.}

\item{clusters}{A boolean indicating whether estimated clusters should be
returned.}

\item{...}{Currently unused.}
}
\value{
A numerical vector with predicted values (if \code{clusters = FALSE})
or predicted clusters (if \code{clusters = FALSE}).
}
\description{
Prediction method for the K-Conditional-Means estimator.
}
\examples{
# Use data from the included simulated dataset
y <- SimDat$y
X <- cbind(SimDat$Z, SimDat$X1, SimDat$X2) # first feature is categorical
# Compute kcmeans with three support points
kcmeans_fit <- kcmeans(y, X, K = 3)
# Calculate in-sample predictions
fitted_values <- predict(kcmeans_fit, X)
# Print sample share of estimated clusters
clusters <- predict(kcmeans_fit, X, clusters = TRUE)
table(clusters)
}
\references{
Wiemann T (2023). "Optimal Categorical Instruments."
}
