% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimDat.R
\docType{data}
\name{SimDat}
\alias{SimDat}
\title{Simulated data.}
\format{
A data frame with 500 rows and 4 variables.
\describe{
\item{y}{The outcome.}
\item{Z0}{The (in practice unobserved) low-dimensional categorical
predictor.}
\item{Z}{The (in practice observed) higher-dimensional categorical
predictor.}
\item{X}{A binary control variable.}
}
}
\usage{
SimDat
}
\description{
Simulated data.
}
\references{
Wiemann (2023). "Optimal Categorical Instruments."
}
\keyword{datasets}
