% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_fun.R
\name{toy_fun}
\alias{toy_fun}
\title{Toy function.}
\usage{
toy_fun(y)
}
\arguments{
\item{x}{A numerical vector.}
}
\value{
\code{toy_fun} returns an object of S3 class
\code{toy_fun}. An object of class \code{toy_fun} is a list containing
the following components:
\describe{
\item{\code{fun}}{A boolean on whether you had fun..}
\item{\code{y}}{Pass-through of the ser-provided arguments.
See above.}
}
}
\description{
This is a simple toy function.
}
\examples{
res <- toy_fun(rnorm(100))
res$fun
}
\concept{toys}
